플레이어 체력회복 로직 해결, 플레이어 피격 로직 구현
플레이어가 회복할 때 재생되는 애니메이션이 움직임으로 인해 캔슬될 수 있도록 구현했으며, 회복 애니메이션이 모두 재생된 후에야 플레이어의 체력이 회복되도록 로직을 완성했습니다. 또한, 몬스터에 피격될 때 피격 애니메이션이 재생되도록 구현했습니다.

+++ Assets/Script/GameManager/GameManager.cs
@@ -4,7 +4,7 @@ public class GameManager : MonoBehaviour
 {
     [Header("플레이어 기본 체력")]
     public int playerMaxHealth = 100; // 플레이어의 초기 체력
-    public int playerHealth; // 플레이어의 현재 체력
+    public float playerHealth; // 플레이어의 현재 체력
 
     [Header("플레이어 기본 공격력")]
     public int playerAttackPower = 10; // 플레이어의 초기 공격력
